move lib functions to lib.py
